requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==5.0.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
executing==2.0.1
hiredis==2.3.2
ipython==8.20.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
python-slugify==8.0.2
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.1
tzdata==2023.4
wcwidth==0.2.13
